Restore Market Lifeline and build Power Relay

+++ Assets/Game/Scripts/UI/Contracts/UiCampaignMissionModels.cs
@@ -95,7 +95,7 @@ namespace Game.UI.Contracts
     {
         public UiCampaignOperationsModel(
             uint version, uint catalogSourceVersion, uint progressSourceVersion,
-            UiCampaignMissionModel selectedMission, string nextMissionId, bool nextMissionRevealed, byte availableMissionMask = 0, byte completedMissionMask = 0)
+            UiCampaignMissionModel selectedMission, string nextMissionId, bool nextMissionRevealed, ushort availableMissionMask = 0, ushort completedMissionMask = 0)
         {
             Version = version;
             CatalogSourceVersion = catalogSourceVersion;
@@ -113,8 +113,8 @@ namespace Game.UI.Contracts
         public UiCampaignMissionModel SelectedMission { get; }
         public string NextMissionId { get; }
         public bool NextMissionRevealed { get; }
-        public byte AvailableMissionMask { get; }
-        public byte CompletedMissionMask { get; }
+        public ushort AvailableMissionMask { get; }
+        public ushort CompletedMissionMask { get; }
         public bool IsValid => Version != 0 && !string.IsNullOrWhiteSpace(SelectedMission.MissionId);
     }
