+++ Assets/Game/Scripts/UI/Shell/Ecs/Contracts/UiCampaignMissionComponents.cs
@@ -25,8 +25,8 @@ namespace Game.UI.Shell.Contracts.Ecs
         public byte FirstClearCompleted;
         public byte PendingResume;
         public byte NextMissionRevealed;
-        public byte AvailableMissionMask;
-        public byte CompletedMissionMask;
+        public ushort AvailableMissionMask;
+        public ushort CompletedMissionMask;
     }
 
     public struct UiMissionObjectiveProjectionData
